fix request answer, and separated api in hpp
