+++ code/own_algorithm/header.hpp
@@ -65,5 +65,6 @@ const int INF = numeric_limits<int>::max();
 #include "landmarks.hpp"
 #include "graph.hpp"
 #include "input.hpp"
+#include "api.hpp"
 
 #endif
